scientific-skills/hypogenic/scripts/requirements.txt
# Hypogenic 依赖
# 安装命令: pip install -r requirements.txt

hypogenic>=0.1.0
pyyaml>=6.0
numpy>=1.20.0
pandas>=1.5.0

# 可选依赖
# redis>=4.5.0  # 用于缓存
# openai>=1.0.0  # 用于 OpenAI API
# anthropic>=0.8.0  # 用于 Anthropic API
